# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulp API",
    "description": "Pulp social-listening HTTP API. This spec is generated from the Rust backend and is the authoritative API contract.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/mentions": {
      "get": {
        "tags": [
          "mentions"
        ],
        "summary": "`GET /api/mentions` — list mentions (keyset-paginated by the compound\n`before`/`before_id` cursor).",
        "operationId": "listMentions",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "monitor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Upper bound (exclusive) on the effective timestamp (`published_at`,\nor `ingested_at` when `published_at` is null). For the feed's\nkeyset-pagination cursor, pair this with `before_id` — the `id` of\nthe last item on the previous page — so rows tied on the effective\ntimestamp are still visited exactly once when paging one page at a\ntime; `before` alone has no tiebreak.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "description": "Tiebreaker id paired with `before` for keyset pagination — the `id`\nof the last item on the previous page. See `before`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Lower bound: return mentions published at or after this epoch.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "read",
            "in": "query",
            "description": "Filter by read state: `true` = read only, `false` = unread only,\nomitted = both.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "ai",
            "in": "query",
            "description": "AI verdict filter: omitted/`visible` = feed default (no verdict or\n`accepted`); `all` = everything; `pending` / `accepted` / `rejected`\n= exactly that verdict.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentionPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/mentions/pending-count": {
      "get": {
        "tags": [
          "mentions"
        ],
        "summary": "`GET /api/mentions/pending-count` — size + age of the AI-filter backlog,\ndriving the feed's \"N pending AI filter\" banner.",
        "operationId": "mentionsPendingCount",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "description": "Scope the count to a single workspace; omitted = all workspaces.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingCount"
                }
              }
            }
          }
        }
      }
    },
    "/api/mentions/{id}": {
      "get": {
        "tags": [
          "mentions"
        ],
        "summary": "`GET /api/mentions/:id` — fetch a single mention (used by the mention\ndetail page that web-push notifications deep-link to).",
        "operationId": "getMention",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Mention id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mention"
                }
              }
            }
          },
          "404": {
            "description": "Mention not found"
          }
        }
      }
    },
    "/api/mentions/{id}/read": {
      "put": {
        "tags": [
          "mentions"
        ],
        "summary": "`PUT /api/mentions/:id/read` — mark a mention read or unread.",
        "operationId": "setMentionRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Mention id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mention"
                }
              }
            }
          },
          "404": {
            "description": "Mention not found"
          }
        }
      }
    },
    "/api/monitors": {
      "get": {
        "tags": [
          "monitors"
        ],
        "summary": "`GET /api/monitors?workspace_id=…` — list monitors in a workspace.\nThe `workspace_id` query param is required.",
        "operationId": "listMonitors",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Monitor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "monitors"
        ],
        "summary": "`POST /api/monitors` — create a monitor watch. Body is [`CreateMonitor`]:\n`workspace_id` and `terms` (the match-any keyword list) are required;\n`channels` (empty/omitted = all channels), `exact_match`, `case_sensitive`,\n`exclude_terms` are optional.",
        "operationId": "createMonitor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonitor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/{id}": {
      "get": {
        "tags": [
          "monitors"
        ],
        "summary": "`GET /api/monitors/:id` — fetch a single monitor.",
        "operationId": "getMonitor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Monitor id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "404": {
            "description": "Monitor not found"
          }
        }
      },
      "put": {
        "tags": [
          "monitors"
        ],
        "summary": "`PUT /api/monitors/:id` — update a monitor watch. Body is [`UpdateMonitor`];\nevery field is optional and only provided fields are changed.",
        "operationId": "updateMonitor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Monitor id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonitor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "monitors"
        ],
        "summary": "`DELETE /api/monitors/:id` — delete a monitor watch (cascades to its mentions).",
        "operationId": "deleteMonitor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Monitor id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMonitor": {
        "type": "object",
        "required": [
          "workspace_id",
          "terms"
        ],
        "properties": {
          "ai_filter_prompt": {
            "type": [
              "string",
              "null"
            ],
            "description": "AI relevance prompt; empty/whitespace is normalized to \"disabled\"."
          },
          "case_sensitive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "channel_settings": {
            "description": "Per-channel collection overrides keyed by channel name."
          },
          "channels": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "exact_match": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "exclude_terms": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Match-any term list (each a bare literal phrase). A post matches if it\ncontains ANY term and none of `exclude_terms`."
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
...
      "Mention": {
        "type": "object",
        "required": [
          "id",
          "monitor_id",
          "channel",
          "external_id",
          "content_text",
          "content_url",
          "ingested_at",
          "platform_meta"
        ],
        "properties": {
          "ai_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "The judge's one-sentence explanation for the verdict."
          },
          "ai_verdict": {
            "type": [
              "string",
              "null"
            ],
            "description": "AI filter verdict: `None` = no AI filtering applied, `pending` =\nawaiting judgment (hidden from the feed), `accepted` / `rejected`."
          },
          "author_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "author_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "channel": {
            "type": "string"
          },
          "content_text": {
            "type": "string"
          },
          "content_url": {
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ingested_at": {
            "type": "integer",
            "format": "int64"
          },
          "monitor_id": {
            "type": "string"
          },
          "platform_meta": {},
          "published_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "read_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "When the user marked this mention read; `None` = unread."
          }
        }
      },
      "MentionPage": {
        "type": "object",
        "description": "A page of mentions plus a cursor hint for keyset pagination.",
        "required": [
          "items",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mention"
            }
          }
        }
      },
...
      "Monitor": {
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "terms",
          "active",
          "channels",
          "exact_match",
          "case_sensitive",
          "exclude_terms",
          "channel_settings",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "ai_filter_prompt": {
            "type": [
              "string",
              "null"
            ],
            "description": "When set, this monitor's new mentions are held out of the feed until\nthe AI judge accepts them against this prompt."
          },
          "case_sensitive": {
            "type": "boolean"
          },
          "channel_settings": {
            "description": "Per-channel collection overrides, keyed by channel name (e.g.\n`{\"reddit\": {\"subreddits\": [\"accessibility\"]}}`). Shallow-merged over\nthe channel's global credentials at collection time; monitor keys win."
          },
          "channels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "exact_match": {
            "type": "boolean"
          },
          "exclude_terms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Match-any keyword/phrase list: a post matches this monitor if it contains\nANY of these terms (per `exact_match` / `case_sensitive`) and none of\n`exclude_terms`. Each term is a bare literal phrase — never OR/quotes."
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
...
      "PendingCount": {
        "type": "object",
        "description": "Size and age of the AI-filter backlog for a workspace — how many mentions\nare still awaiting a verdict (withheld from the feed) and when the oldest was\ningested. The feed shows a \"N pending AI filter\" banner once the oldest has\naged past a staleness threshold.",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "oldest_ingested_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "`ingested_at` (epoch seconds) of the oldest pending mention, or `null`\nwhen nothing is pending."
          }
        }
      },
      "SetReadRequest": {
        "type": "object",
        "description": "Body for [`set_read`].",
        "required": [
          "read"
        ],
        "properties": {
          "read": {
            "type": "boolean",
            "description": "`true` marks the mention read, `false` marks it unread."
          }
        }
      },
...
      "UpdateMonitor": {
        "type": "object",
        "properties": {
          "active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "ai_filter_prompt": {
            "type": [
              "string",
              "null"
            ],
            "description": "AI relevance prompt; pass an empty string to clear."
          },
          "case_sensitive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "channel_settings": {
            "description": "Per-channel collection overrides; pass `{}` to clear."
          },
          "channels": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "exact_match": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "exclude_terms": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "terms": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Replace the match-any term list. Omit to leave it unchanged."
          }
        }
      },
...
    }
  },
...
}